% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gest}
\alias{gest}
\title{Number of gestures by infants at 10, 11 and 12 months}
\format{
A tibble with 540 observations and 5 variables: \describe{
\item{dyad}{Unique parent/infant dyad ID.}
\item{background}{Cultural background of dyad.}
\item{months}{Time point in infant months.}
\item{gesture}{Type of gesture.}
\item{count}{Number of gestures.}
}
}
\source{
\doi{10.1111/cdev.13406}
}
\usage{
gest
}
\description{
This data table contains counts of three type of gestures performed by 60
infants  from Bengali, Chinese and British backgrounds.
}
\keyword{datasets}
