% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{struct}
\alias{struct}
\title{ERP to structural violation in music and language}
\format{
A tibble with 17160 observations and 6 variables: \describe{
\item{t}{Time from stimulus onset in milliseconds.}
\item{electrode}{Electrode number.}
\item{voltage}{Electrode voltage at time t.}
\item{stimulus.condition}{Language vs music.}
\item{grammar.condition}{Structural type (grammatical vs ungrammatical).}
}
}
\source{
\url{https://doi.org/10.31234/osf.io/e9w3v}
}
\usage{
struct
}
\description{
This data table contains ERP amplitude data from 39 subjects listening to speech and music.
}
\keyword{datasets}
